backend/pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[dependency-groups]
dev = [
    "coverage>=7.6.10",
    "mdformat-footnote>=0.1.1",
    "mdformat-frontmatter>=2.0.8",
    "mdformat-gfm-alerts>=1.0.1",
    "mdformat>=0.7.21",
    "mypy>=1.8.0",
    "pre-commit>=3.7.0",
    "ruff>=0.4.3",
    "rust-just>=1.39.0",
    "gradescopeapi",
]

[project]
authors = [
    { name = "Berry Liu" },
    { name = "Calvin Tian" },
    { name = "Kevin Zheng" },
    { name = "Margaret Jagger" },
    { name = "Susmitha Kusuma" },
]
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "beautifulsoup4>=4.12.3",
    "fastapi>=0.111.0",
    "pytest>=8.2.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "requests-toolbelt>=1.0.0",
    "requests>=2.31.0",
    "tzdata>=2024.2",
]
description = "Library for programmatically interacting with Gradescope."
maintainers = [
    { name = "Calvin Tian" },
]
name = "gradescopeapi"
readme = "README.md"
requires-python = ">=3.10"
version = "1.4.1"

[project.license]
text = "MIT"

[project.urls]
Homepage = "https://github.com/nyuoss/gradescope-api"
Issues = "https://github.com/nyuoss/gradescope-api/issues"
Repository = "https://github.com/nyuoss/gradescope-api"

[tool.uv.sources]
gradescopeapi = { workspace = true }
